+++ src/lib/presets/vanilla.js
@@ -3,11 +3,13 @@ import List from '../List.svelte';
 import Item from '../Item.svelte';
 import Selection from '../Selection.svelte';
 import debounce from '../debounce';
+import filter from '../filter';
 
 export default {
     theme: defaultStyles,
     List,
     Item,
     Selection,
-    debounce
+    debounce,
+    filter
 }