+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## unreleased
 
+* filterMethod changed to filter and added via config
 * added support for svelte-tiny-virtual-list
 * removed virtual-list class and css props
 * moved VirtualList to config